chore: float->double because neo4j support is crap

+++ core/src/main/kotlin/terapps/factoryplanner/core/entities/Automaton.kt
@@ -2,7 +2,7 @@ package terapps.factoryplanner.core.entities
 
 interface Automaton {
     val className: String
-    val minPotential: Float
-    val maxPotential: Float
-    val productionBoost: Float
+    val minPotential: Double
+    val maxPotential: Double
+    val productionBoost: Double
 }
