+++ core/src/main/kotlin/terapps/factoryplanner/core/projections/ItemIO.kt
@@ -2,5 +2,5 @@ package terapps.factoryplanner.core.projections
 
 interface ItemIO {
     fun getDescriptor(): ItemDescriptorSummary
-    fun getOutputPerCycle(): Float
+    fun getOutputPerCycle(): Double
 }
