+++ core/src/main/kotlin/terapps/factoryplanner/core/services/components/FactorySite.kt
@@ -4,14 +4,14 @@ import terapps.factoryplanner.core.entities.Automaton
 import terapps.factoryplanner.core.projections.ItemDescriptorSummary
 data class FactorySiteIO(
         val item: ItemDescriptorSummary,
-        val outputPerCycle: Float
+        val outputPerCycle: Double
 )
 
 abstract class FactorySite {
     abstract val targetDescriptor: ItemDescriptorSummary
-    abstract var targetAmountPerCycle: Float
+    abstract var targetAmountPerCycle: Double
     abstract val produces: List<FactorySiteIO>
-    abstract val requiredMachines: Float
+    abstract val requiredMachines: Double
     abstract val automaton: Automaton
     var destinations: MutableList<Destination> = mutableListOf()
     val childs: MutableList<FactorySite> = mutableListOf()
